+++ packages/pronotron-pointer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pronotron/io
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pronotron/utils@1.1.0
+
 ## 1.0.1
 
 ### Patch Changes